Move CLI logic into cli submodule
